Use 1-AUC as validation loss for hyperparameter search
During cross-validation for hyperparameter search, this was using the
same loss function as used during training (binary cross-entropy) to
evaluate the performance of the model on the validation set. Now, this
uses 1-AUC as the loss function to minimize -- ultimately, in evaluating
the model, we are looking at AUC. Here, AUC is from the ROC curve.