Improve baseline classification models
This revamps predictor_baseline in several ways. It uses sklearn
for logistic regression instead of TensorFlow. It uses both L1 and
L2 logistic regression, and uses sklearn's built-in hyperparameter
search via cross-validation to determine the best hyperparameters (here,
just the regularization strength). This should improve the performance
of the baseline classifier.

The commit also generalizes predictor_baseline to make it easier to
include regression models.