Add option to explicitly add GC content as a feature to the predictor's fully connected layer
GC content of a guide appears to be a variable significantly
(negatively) correlated with a guide's activity. The predictor can learn
this feature on its own, but providing it 'manually' into the network
should help it avoid having to learn this and perhaps to a better job
constructing representations of other features.