Save all models during hyperparameter search and skip choosing a 'best'
This changes the hyperparameter search to save all model hyperparameters
and trained weights, and still output all metrics across all folds.
While it does still impose a criterion to choose a 'best' model, it does
not run it on the test data; predictor_hyperparam_search.py should no
longer use the test data. predictor.py can load a model and run it on
the test data.