Add option to normalize activity of each crRNA
This adds an argument, `--cas13-normalize-crrna-activity`, which
normalizes the activity of each Cas13 guide/crRNA across its targets to
have mean of 0 and standard deviation of 1. This enables models to
predict the effects of differences in targets (e.g., owing to
mismatches) rather than differences owing to something (e.g., sequence
composition) inherent in the crRNAs.