Add results on test set for best model from normalized crRNA hyperparameter search
This model is for the Cas13 'exp-and-pos' dataset, regressing on active
data points, where the training and prediction is done such that the
activity is normalized across targets for each crRNA. The hyperparameter
search results are in commit 0cb2c27.

This commit includes results of running predictor.py on the selected
model, and plots of the results.