Add function to weight samples in predictor during regression
This adds the function sample_regression_weight() in the parse_data
module. This is used by predictor. It weights samples according to the
z-score within a guide -- i.e., so that guide-target pairs where the
activity of a target deviates from the mean for the guide (across
targets) are weighted more heavily. This is controlled by a parameter p,
currently set to 0 (i.e., so samples are not weighted unless p is set to
be nonzero).