Use replicate-level measurements for Cas13 guide-target pair data
This randomly samples the same number of replicate measurements (10) for
each guide-target pair. Therefore, if the number of guide-target pairs
was N, we'll wind up with 10*N data points (or 'samples') in our
dataset.

A choice of 10 is ok because most guide-target pairs have >=10 technical
replicate measurements from which to sample.