Remove Cas9 and simulated Cas13 boilerplate code
The Cas9 data was from Doench et al., and the simulated Cas13 data was
simulated from that Cas9 data.

This removes the code that parses the data (in parse_data.py), as well
as arguments for reading/using it. It also removes the output/results on
this data. It moves the data into a cas9/ subfolder in data/.

The purpose is just to cleanup this repo since we are no longer using
that dataset.