Stop using global variables in parse_data module
The parse_data module used several global variables (*_parser) to store
references to an activity data parser object from that module. This was
messy.

This commit changes the functions that require those variables to be
methods inside the activity data parser class.

This also removes another global variable, storing context_nt, and
changes the function using it to accept that value as an argument.