Do not use the 'add_gc_content' or 'sample_weight_scaling_factor' hyperparameters
This leaves them implemented, but changes the hyperparameter search to
set them to not be used. Neither appears to have much of an impact,
according to a plot of losses against different hyperparameters.  Also,
the 'sample_weight_scaling_factor' parameter is no longer used because
the train_with_keras and test_with_keras functions do not use
sample_weight for regressions.