Fix bug when determining active data points
This had defined active data points to be `>= ACTIVITY_THRESHOLD`.
However, when the ACTIVITY_THRESHOLD was set to -4.0 (the lowest
activity in the measurements), this would cause all data points to be
active. As a result, when using --cas13-regress-only-on-active, there
would be no filtering -- even the data points that were supposed to be
negative (-4.0) would be included. This commit fixes this bug.

Note that classifying activity for --cas13-classify was performed
correctly (activity `<= ACTIVITY_THRESHOLD` was inactive, and otherwise
active).