Serialize Cas13 classify/regress models with different versions of TensorFlow
This commit adds serialized models using TensorFlow 2.1, 2.2, and 2.3.
Even though TensorFlow claims to support forward compatibility, the
original 2.1-serialized model cannot be read by 2.2 or 2.3 (possibly due
to changes to SavedModel made in 2.2).

Here is a table in which columns give the TensorFlow version used to
serialize the model and the rows give a version used to try to load the
model (Yes=loads successfully, No=does not):
         Serialized with:
         2.1   2.2   2.3
Loads
with:
  2.1    Yes   Yes   No
  2.2    No    Yes   No
  2.3    No    Yes   Yes