Compute classifier confusion matrix across folds and at different thresholds
This commit modifies predictor.py to have the
--compute-confusion-matrices-across-thresholds-and-folds argument, which
computes the number of true/false negatives/positives at each threshold
across 5 folds of the training data (it does not use the test data), and
writes them to a file.

Note that some values in test.tsv.gz changed very slightly despite using
the same seed; I think this is due to rounding differences compared to
the last run, possibly due to different versions of NumPy/TensorFlow.
Likewise the optimal threshold for a 0.975 threshold, in test.out.gz, is
slightly different.

This commit also includes the new outputs.