Add option to stop early the evaluation of a hyperparameter choice
This halts after evaluating an outer fold whose loss falls below a
specified value. It is helpful if we know a range of "good" values
because it can significantly speed runtime of the search to avoid
fully evaluating (across all outer folds) a choice of hyperparameters
that is unlikely to be the best. However, with its use, the results
cannot be used to evaluate the significance of a hyperparameter based on
a marginal distribution.