Add analysis of effect of locally connected layer
This performs a nested cross-validation with and without requiring a
locally-connected (LC) layer in the model. The resulting values (on the
held-out fold) can inform whether the LC layer improves model
performance. This is a better way of testing the effect of the LC layer
than looking at the whole hyperparameter search (comparing choices with
the LC layer vs. choice without) -- e.g., there may be a good model
selection without an LC layer that the hyperparameter search cannot
find.

This commit also contains the output of the analyses.