Print summary statistics for classifier performance on subsets of test data
This prints AUC for the ROC and PR curves when the classifier is run
only on test data with certain Hamming distances and PFSs. This provides
performance, for example, of how well the classifier model does only on
perfect matching guide-target pairs.

The output PDFs from the script should not have changed. This commit
also includes the output (plot.out.gz) of running the plotting script.