Make plots of regression results on subsets of test data without including measurement error in the true activities
Several previous commits created some regression result plots without
including measurement error (i.e., replicate measurements) in the plots
-- in the scatter plots, these create a straight line for each
guide-target pair if true activity is on the horizontal axis. We can use
the mean of the replicate measurements as a single value for true
activity. This commit updates the plotting script to also exclude the
visualization of measurement error (and exclude from the correlation
statistics) in plots that show the model's results on subsets of the
test data.

The commit also includes the PDF outputs of running the plotting script,
both for test data that is true active and classified to be active.