Fix bug when plotting regression results without measurement error
To remove measurement error, the plotting script changed the true (and
predicted) activity value for each guide-target pair to be the mean
across technical replicates. But it did not collapse down those
replicates to a single pair -- so the same number of data points
remained (i.e., with all technical replicates) but just the replicates
shared the same (mean) activity value for each guide-target pair. This
could affect the plot and the metrics (e.g., Spearman correlation) in a
minor way because the number of replicates may not be the exactly the
same for each guide-target pair: though it is usually 10 (after
resampling), this script only considers active replicates and some
replicates for a guide-target pair may be inactive while others are
active.

This commit fixes the issue by collapsing the replicates for each
guide-target pair, and taking the mean across the measurements.